fix: コンポーネント名をTextInputに変更, labelをコンポーネント内から削除
